CLJS-519: namespace aliased keywords throw Clojure reader exception
`forms-seq` was not as lazy at it should have been, we might
accidentally look ahead past the ns form and consume a namespace
aliased keyword before we get a chance to apply the side effects of
the `parse 'ns` multimethod in the analyzer.

We were incorrectly setting `*ns*` in `forms-seq` to
`*ana/*reader-ns*` instead of `*ana/cljs-ns*` which is what the analyzer
actually side effects.